starbucks_fastapi_classical_individual/requirements.txt
fastapi
uvicorn
pydantic
tensorflow
pandas
numpy
scikit-learn
yfinance
pydot
graphviz
matplotlib
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-multipart
PyJWT
